chore: organize react components
