+++ assets/src/apps/bus_eink.tsx
@@ -1,6 +1,6 @@
 declare function require(name: string): string;
 // tslint:disable-next-line
-require("../css/gl_eink_double.scss");
+require("../../css/bus_eink.scss");
 
 import React from "react";
 import ReactDOM from "react-dom";
@@ -10,16 +10,16 @@ import {
   AuditScreenPage,
   MultiScreenPage,
   ScreenPage
-} from "Components/eink/green_line/double/screen_page";
+} from "Components/eink/bus/screen_page";
 
 const App = (): JSX.Element => {
   return (
     <Router>
       <Switch>
-        <Route exact path="/screen/gl_eink_double">
+        <Route exact path="/screen/bus_eink">
           <MultiScreenPage />
         </Route>
-        <Route exact path="/audit/gl_eink_double">
+        <Route exact path="/audit/bus_eink">
           <AuditScreenPage />
         </Route>
         <Route path="/screen/:id">
