+++ assets/src/apps/gl_eink_single.tsx
@@ -1,6 +1,6 @@
 declare function require(name: string): string;
 // tslint:disable-next-line
-require("../css/gl_eink_single.scss");
+require("../../css/gl_eink_single.scss");
 
 import React from "react";
 import ReactDOM from "react-dom";
