+++ assets/src/components/eink/bus/departure_route.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import { classWithSize } from "../util";
+import { classWithSize } from "Util";
 
 const DepartureRoute = ({ route, size }): JSX.Element => {
   if (!route) {
