+++ assets/src/components/eink/bus/departures.tsx
@@ -1,4 +1,4 @@
-import DeparturesRow from "Components/departures_row";
+import DeparturesRow from "Components/eink/bus/departures_row";
 import React, { forwardRef } from "react";
 
 const buildDeparturesRows = departures => {
