+++ assets/src/components/eink/bus/departures_row.tsx
@@ -1,8 +1,8 @@
 import React from "react";
 
+import DepartureRow from "Components/eink/bus/departure_row";
 import InlineAlert from "Components/eink/bus/inline_alert";
-import { classWithSize } from "../util";
-import DepartureRow from "./departure_row";
+import { classWithSize } from "Util";
 
 const DeparturesRow = ({
   currentTimeString,
