+++ assets/src/components/eink/connection_error.tsx
