+++ assets/src/components/eink/digital_bridge.tsx
