+++ assets/src/components/eink/global_alert.tsx
