+++ assets/src/components/eink/green_line/nearby_departures.tsx
