+++ assets/src/components/eink/green_line/single/screen_container.tsx
@@ -1,12 +1,12 @@
 import React, { useEffect, useState } from "react";
 
-import ConnectionError from "Components/connection_error";
-import DigitalBridge from "Components/digital_bridge";
+import ConnectionError from "Components/eink/connection_error";
+import DigitalBridge from "Components/eink/digital_bridge";
 import Departures from "Components/eink/green_line/departures";
 import Header from "Components/eink/green_line/header";
 import LineMap from "Components/eink/green_line/line_map";
-import { NoServiceTop } from "Components/no_service";
-import OvernightDepartures from "Components/overnight_departures";
+import { NoServiceTop } from "Components/eink/no_service";
+import OvernightDepartures from "Components/eink/overnight_departures";
 
 import useApiResponse from "Hooks/use_api_response";
 
@@ -68,8 +68,9 @@ const NoServiceScreenLayout = (): JSX.Element => {
 };
 
 const NoDeparturesScreenLayout = ({ apiResponse }): JSX.Element => {
-  // We successfully fetched data, but there are no predictions.
-  // For now, assume that this is because it's the middle of the night.
+  // We successfully fetched data, but there are no predictions, and we don't have
+  // a headway for the current daypart. For now, we assume that it's the middle of
+  // the night.
   return (
     <OvernightDepartures
       size="single"
