+++ assets/src/components/eink/no_service.tsx
@@ -198,3 +198,4 @@ const NoService = ({ mode }): JSX.Element => {
 };
 
 export { NoServiceTop, NoServiceBottom };
+export default NoService;
