+++ assets/src/components/eink/takeover_alert.tsx
